--- methods/GSGraphQLBackend/graphql/package.json
@@ -1,74 +1,0 @@
-{
-  "name": "@groceristar/data-methods-graphql",
-  "version": "0.0.1",
-  "license": "MIT",
-  "scripts": {
-    "bundle": "rollup -c",
-    "dev": "rollup -c -w",
-    "build": "npm run clean && npm run bundle",
-    "build:babel": "babel src -d dist",
-    "clean": "shx rm -rf ./coverage && shx rm -rf ./dist",
-    "test": "jest",
-    "test:watch": "jest --watch",
-    "lint": "eslint ./src --fix --ignore-pattern \"./package.json\" ",
-    "lint-staged": "lint-staged -c lint-staged.config.js",
-    "format": "prettier ./src/** --write",
-    "formato": "pretty-quick --staged"
-  },
-  "description": "module to use static JSON files with groceristar projects",
-  "main": "dist/index.cjs.js",
-  "module": "dist/index.es.js",
-  "browser": "dist/index.iife.js",
-  "repository": "https://github.com/GroceriStar/food-static-files-generator",
-  "author": "Arthur Tkachenko",
-  "private": false,
-  "release": {
-    "publishDir": "dist"
-  },
-  "devDependencies": {
-    "@babel/cli": "7.6.2",
-    "@babel/core": "7.6.2",
-    "@babel/plugin-proposal-object-rest-spread": "7.6.2",
-    "@babel/plugin-syntax-dynamic-import": "7.2.0",
-    "@babel/plugin-transform-async-to-generator": "7.5.0",
-    "@babel/plugin-transform-runtime": "7.6.2",
-    "@babel/preset-env": "7.6.2",
-    "babel-core": "7.0.0-bridge.0",
-    "babel-eslint": "10.0.3",
-    "babel-jest": "24.9.0",
-    "babel-plugin-module-resolver": "3.2.0",
-    "coveralls": "3.0.6",
-    "eslint": "6.5.0",
-    "eslint-config-airbnb-base": "14.0.0",
-    "eslint-plugin-import": "2.18.2",
-    "eslint-plugin-jest": "22.17.0",
-    "husky": "3.0.7",
-    "jest": "24.9.0",
-    "lint-staged": "9.4.0",
-    "prettier": "1.18.2",
-    "pretty-quick": "1.11.1",
-    "regenerator-runtime": "0.13.3",
-    "rollup": "1.22.0",
-    "rollup-plugin-babel": "4.3.3",
-    "rollup-plugin-commonjs": "10.1.0",
-    "rollup-plugin-eslint": "7.0.0",
-    "rollup-plugin-node-builtins": "2.1.2",
-    "rollup-plugin-node-resolve": "5.2.0",
-    "rollup-plugin-replace": "2.2.0",
-    "shx": "0.3.2"
-  },
-  "peerDependencies": {},
-  "publishConfig": {
-    "access": "public"
-  },
-  "dependencies": {
-    "dayjs": "^1.8.13",
-    "fs": "^0.0.2",
-    "is-valid-path": "^0.1.1",
-    "lodash": "^4.17.11",
-    "path": "^0.12.7"
-  },
-  "files": [
-    "dist"
-  ]
-}
