--- methods/GSGraphQLBackend/graphql/renovate.json
@@ -1,5 +1,0 @@
-{
-  "extends": [
-    "config:base"
-  ]
-}
