+++ methods/GSGraphQLBackend/rollup.config.js
@@ -1,110 +1,141 @@
-// plugins that we are going to use
-import commonjs from 'rollup-plugin-commonjs'
-import resolve from 'rollup-plugin-node-resolve'
-import json from 'rollup-plugin-json'
-// import copy from 'rollup-plugin-copy-assets'
-import builtins from 'rollup-plugin-node-builtins'
-import babel from 'rollup-plugin-babel'
-import pkg from './package.json'
-import notify from 'rollup-plugin-notify'
+import commonjs from 'rollup-plugin-commonjs';
+import resolve from 'rollup-plugin-node-resolve';
+import builtins from 'rollup-plugin-node-builtins';
+import replace from 'rollup-plugin-replace';
+import babel from 'rollup-plugin-babel';
+// import eslint from 'rollup-plugin-eslint';
+import pkg from './package.json';
 
-// import cleanup from 'rollup-plugin-cleanup';
-// https://github.com/mjeanroy/rollup-plugin-prettier
-// https://gitlab.com/IvanSanchez/rollup-plugin-file-as-blob
+// not all files you want to resolve are .js files
+// Default: [ '.mjs', '.js', '.json', '.node' ]
+const extensions = ['.js'];
 
-const extensions = [
-  '.js', '.json'
-]
+const name = 'StaticFilesGenerator';
 
-const name = 'StaticDataWrapper'
+// packages that should be treated as external dependencies, not bundled
+// e.g. ['axios']
+const external = [
+  'lodash',
+  '@groceristar/static-data-generator',
+  '@groceristar/sd-wrapper',
+  'uuid/v1',
+];
 
-// packages that should be treated as external or global dependencies, not bundled
-const { external, globals } = {
-  'globals': {
-    'fs': 'fs',
-    'uuid': 'uuid',
-    'path': 'path'
-  },
-  'external': [
-    'fs',
-    'path',
-    'uuid'
-  ]
-}
+// list of plugins used during building process
+const plugins = () => [
+  // Allows node_modules resolution
+  resolve({
+    extensions,
+
+    // the fields to scan in a package.json to determine the entry point
+    // if this list contains "browser", overrides specified in "pkg.browser"
+    // will be used
+    mainFields: ['module', 'main', 'browser'], // Default: ['module', 'main']
+  }),
+
+  // @TODO maybe we should do it for production only? not sure
+  replace({
+    include: ['node_modules/uuid/**'],
+    delimiters: ['', ''],
+    values: {
+      'crypto.randomBytes': 'require(\'randombytes\')',
+      //   "__BUILD_DATE__": () => new Date().toISOString(),
+      //   "__VERSION__": fs.readFileSync("version", "utf8").trim()
+    },
+  }),
+
+  // Allows verification of entry point and all imported files with ESLint.
+  // @TODO fix and enable eslint for rollup
+  // eslint({
+  //   /* your options */
+  //   fix:true,
+  //   throwOnWarning:true,
+  //   throwOnError:true
+
+  // }),
+
+  // Allow bundling cjs modules. Rollup doesn't understand cjs
+  commonjs({
+    ignore: ['conditional-runtime-dependency'],
+  }),
+
+  // use Babel to compile TypeScript/JavaScript files to ES5
+  babel({
+    extensions,
+    include: ['src/*'],
+    // ignore node_modules/ in transpilation process
+    exclude: 'node_modules/**',
+    // ignore .babelrc (if defined) and use options defined here
+    // babelrc: false,
+    // use recommended babel-preset-env without es modules enabled
+    // and with possibility to set custom targets e.g. { node: '8' }
+    // presets: [['env', { modules: false, targets }]],
+    // solve a problem with spread operator transpilation https://github.com/rollup/rollup/issues/281
+    // plugins: ['babel-plugin-transform-object-rest-spread'],
+    // removes comments from output
+    comments: false,
+    runtimeHelpers: true,
+  }),
+
+  builtins(),
+  // remove flow annotations from output
+  // flow(),
+
+  // copy Flow definitions from source to destination directory
+  // copy({
+  //   files: ['src/*.flow'],
+  //   dest: 'dist',
+  // }),
+];
+
+// example for adding plugin for env only
+// if(process.env.NODE_ENV == "production") {
+//   console.log("[config] In production environment - minifying JS");
+//   plugins.push(terser({
+//     numWorkers: os.cpus().length,
+//     compress: {
+//       ecma: 6
+//     }
+//   }));
+// }
 
 export default {
   // source file / entrypoint
-  input: './src/index.js',
-
-  // Specify here external modules which you don't want to include in your bundle (for instance: 'lodash', 'moment' etc.)
-  // https://rollupjs.org/guide/en#external-e-external
-  external,
-  globals,
-
-  // list of plugins used during building process
-  plugins: [
-    // Allows node_modules resolution
-    resolve({ extensions }),
-
-    // Allow bundling cjs modules. Rollup doesn't understand cjs
-    commonjs(),
-
-    // Compile TypeScript/JavaScript files
-    babel({
-      extensions,
-      include: ['src/*'],
-      // include: ['src/**/*'],
-      exclude: [
-        'node_modules/**',
-        '/src/data/__tests__'
-        // '/src/data/json-tests'
-      ]
-      // exclude: 'node_modules/**'
-      // presets: presets,
-      // plugins: plugins
-    }),
-
-    // Allow Rollup to import data from JSON file
-    // json()
-    json({
-      include: 'src/data/**',
-
-      // for tree-shaking, properties will be declared as
-      // variables, using either `var` or `const`
-      preferConst: true,
-
-      // generate a named export for every property of the JSON object
-      namedExports: true // Default: true
-
-    }),
-
-    // juck fix in order to move json files to npm
-    // copy({
-    //   assets: [
-    //     './src/data/**'
-    //   ],
-    // }),
-
-    builtins(),
-
-    // Displays rollup errors as system notifications
-    notify()
-
+  input: 'src/index.js',
+  // output configuration
+  output: [
+    {
+      // output file location
+      file: pkg.main,
+      // format of generated JS file, also: esm, and others are available
+      format: 'cjs',
+    },
+    {
+      // output file location
+      file: pkg.module,
+      // format of generated JS file, also: esm, and others are available
+      format: 'es',
+      // format: 'esm',
+      // add sourcemaps
+      sourcemap: true,
+    },
+    {
+      // output file location
+      file: pkg.browser,
+      // format of generated JS file, also: esm, and others are available
+      format: 'iife',
+      // name visible for other scripts
+      name,
+      // https://rollupjs.org/guide/en#output-globals-g-globals
+      // globals: {}
+    },
   ],
 
-  // output configuration
-  output: [{
-    file: pkg.main,
-    format: 'cjs'
-  }, {
-    file: pkg.module,
-    format: 'es'
-  }, {
-    file: pkg.browser,
-    format: 'iife',
-    name
+  // Specify here external modules which you don't want to include
+  // in your bundle (for instance: 'lodash', 'moment' etc.)
+  // https://rollupjs.org/guide/en#external-e-external
+  external,
 
-    // https://rollupjs.org/guide/en#output-globals-g-globals
-    // globals: {}
-  }]
-}
+  // build es modules or commonjs
+  plugins: plugins(),
+};
